+++ pykechain/models/property_reference.py
@@ -3,9 +3,13 @@ from typing import List, Optional
 from pykechain.defaults import PARTS_BATCH_LIMIT
 from pykechain.exceptions import IllegalArgumentError
 from pykechain.models import Activity, Scope, user
-from pykechain.models.base_reference import _ReferenceProperty, _ReferencePropertyInScope
+from pykechain.models.base_reference import (
+    _ReferenceProperty,
+    _ReferencePropertyInScope,
+)
 from pykechain.models.context import Context
 from pykechain.models.form import Form
+from pykechain.models.stored_file import StoredFile
 from pykechain.models.value_filter import ScopeFilter
 from pykechain.models.workflow import Status
 from pykechain.utils import get_in_chunks
@@ -171,7 +175,7 @@ class FormReferencesProperty(_ReferencePropertyInScope):
     .. versionadded:: 3.7
     """
 
-    # REFERENCED_CLASS = Form
+    REFERENCED_CLASS = Form
 
     def _retrieve_objects(self, **kwargs) -> List[Form]:
         """
@@ -194,7 +198,7 @@ class ContextReferencesProperty(_ReferencePropertyInScope):
     .. versionadded:: 3.7
     """
 
-    # REFERENCED_CLASS = Context
+    REFERENCED_CLASS = Context
 
     def _retrieve_objects(self, **kwargs) -> List[Context]:
         """
@@ -212,23 +216,42 @@ class ContextReferencesProperty(_ReferencePropertyInScope):
 
 
 class StatusReferencesProperty(_ReferenceProperty):
-    """A virtual object representing a KE-chain Context References property.
+    """A virtual object representing a KE-chain Status References property.
 
     .. versionadded:: 3.19
     """
 
-    # REFERENCED_CLASS = Status
+    REFERENCED_CLASS = Status
 
-    def _retrieve_objects(self, **kwargs) -> List[Context]:
+    def _retrieve_objects(self, **kwargs) -> List[Status]:
         """
-        Retrieve a list of Contexts.
+        Retrieve a list of Statuses.
 
         :param kwargs: optional inputs
-        :return: list of Context objects
+        :return: list of Status objects
         """
         statuses = []
-        for statuse_json in self._value:
-            status = Status(client=self._client, json=statuse_json)
+        for status_json in self._value:
+            status = Status(client=self._client, json=status_json)
             status.refresh()  # To populate the object with all expected data
             statuses.append(status)
         return statuses
+
+
+class StoredFilesReferencesProperty(_ReferenceProperty):
+    """A virtual object representing a KE-chain StoredFile References property.
+
+    .. versionadded:: 4.7
+    """
+
+    REFERENCED_CLASS = StoredFile
+
+    def _retrieve_objects(self, **kwargs) -> List[StoredFile]:
+        """
+        Retrieve a list of StoredFile.
+
+        :param kwargs: optional inputs
+        :return: list of StoredFile objects
+        """
+        return [StoredFile(client=self._client, json=stored_files_json) for stored_files_json in
+                self._value]